git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3353 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/cam_andor/cam_andor.cpp
@@ -727,8 +727,8 @@ void QFExtensionCameraAndor::changeCameraSetting(QSettings& settings, QFExtensio
     if (which==QFExtensionCamera::CamSetExposureTime) settings.setValue(prefix+"exposure_time", value);
     if (which==QFExtensionCamera::CamSetNumberFrames) settings.setValue(prefix+"kinetic_cycles", value);
     if (which==QFExtensionCamera::CamSetEMGAIN) {
-        settings.setValue(prefix+"emgain", value);
-        settings.setValue(prefix+"emgain_enabled", value>0);
+        settings.setValue(prefix+"emgain", value.toDouble());
+        settings.setValue(prefix+"emgain_enabled", value.toDouble()>0);
     }
     if (which==QFExtensionCamera::CamSetGain) settings.setValue(prefix+"preamp_gain", value);
     if (which==QFExtensionCamera::CamSetFrametime) settings.setValue(prefix+"kinetic_cycle_time", value);
